client: fix sign/unsigned comparison
    /home/pdonnell/ceph/src/client/Client.cc:10906:17: warning: comparison of integer expressions of different signedness: ‘int64_t’ {aka ‘long int’} and ‘uint64_t’ {aka ‘long unsigned int’} [-Wsign-compare]
    10906 |     if ((offset >= in->size) || (size == 0)) {
          |          ~~~~~~~^~~~~~~~~~~

Fixes: https://tracker.ceph.com/issues/66373

+++ src/client/Client.cc
@@ -10991,7 +10991,7 @@ retry:
     // C_Read_Sync_NonBlocking::finish().
 
     // trim read based on file size?
-    if ((offset >= in->size) || (size == 0)) {
+    if (((uint64_t)offset >= in->size) || (size == 0)) {
       // read is requested at the EOF or the read len is zero, therefore just
       // release managed pointers and complete the C_Read_Finisher immediately with 0 bytes
